fix: remove explicit Int64 cast from label column
- Remove .cast(pl.Int64) to match binary version behavior
- Let Polars infer type naturally from NumPy int array
- Fixes CI type incompatibility issue

+++ loop/sfm/lightgbm/utils/tradeline_multiclass.py
@@ -338,7 +338,7 @@ def create_multiclass_labels(df: pl.DataFrame,
                 labels[idx] = 2
                 short_count += 1
 
-    df = df.with_columns(pl.Series('label', labels).cast(pl.Int64))
+    df = df.with_columns(pl.Series('label', labels, dtype=pl.Int32))
 
     return df
